requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.1.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==6.0.0
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.8.0
etelemetry==0.3.1
faiss-cpu==1.7.4
filelock==3.13.1
fonttools==4.47.0
frontend==0.0.3
frozenlist==1.4.0
fsspec==2023.12.2
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
grpcio==1.60.0
grpcio-tools==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
huggingface-hub==0.19.4
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.11.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
langchain==0.0.348
langchain-core==0.0.12
langsmith==0.0.69
llama-index==0.9.16.post1
looseversion==1.3.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
nltk==3.8.1
numpy==1.26.2
openai==1.3.8
packaging==23.2
pandas==2.1.4
pathlib==1.0.1
pdf2image==1.16.3
Pillow==10.1.0
portalocker==2.8.2
protobuf==4.25.1
prov==2.0.0
pyarrow==14.0.1
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pydot==1.4.2
Pygments==2.17.2
pyparsing==3.1.1
pypdf==3.17.2
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytils==0.4.1
pytz==2023.3.post1
pywin32==306
pyxnat==1.6
PyYAML==6.0.1
qdrant-client==1.7.0
rdflib==7.0.0
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
safetensors==0.4.1
scipy==1.11.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.34.0
streamlit==1.29.0
streamlit-chat==0.1.1
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.0
toml==0.10.2
tools==0.1.9
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traits==6.3.2
transformers==4.36.2
types-protobuf==4.24.0.4
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.24.0.post1
validators==0.22.0
watchdog==3.0.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
